refactor code

+++ app/src/main/java/com/swachev/model/Address.kt
@@ -3,7 +3,7 @@ package com.swachev.model
 data class Address(
     val city: String,
     val country: String,
-    val lattitude: Int,
+    val lattitude: Long,
     val line1: String,
     val line2: String,
     val longitude: Long,
